requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
django-quill-editor==0.1.40
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.4.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.1
tzdata==2022.6
tzlocal==4.2
